move test helper functions to the testing file
